--- app/src/main/java/com/belatrixsf/allstars/networking/retrofit/RetrofitFactory.java
@@ -1,83 +1,0 @@
-/* The MIT License (MIT)
-* Copyright (c) 2016 BELATRIX
-* Permission is hereby granted, free of charge, to any person obtaining a copy
-* of this software and associated documentation files (the "Software"), to deal
-* in the Software without restriction, including without limitation the rights
-* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-* copies of the Software, and to permit persons to whom the Software is
-* furnished to do so, subject to the following conditions:
-
-* The above copyright notice and this permission notice shall be included in all
-* copies or substantial portions of the Software.
-
-* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-* SOFTWARE.
-*/
-package com.belatrixsf.allstars.networking.retrofit;
-
-
-import com.belatrixsf.allstars.managers.PreferencesManager;
-import com.squareup.okhttp.Interceptor;
-import com.squareup.okhttp.OkHttpClient;
-import com.squareup.okhttp.Request;
-import com.squareup.okhttp.Response;
-import com.squareup.okhttp.logging.HttpLoggingInterceptor;
-
-
-import java.io.IOException;
-
-import retrofit.GsonConverterFactory;
-import retrofit.Retrofit;
-
-/**
- * Created by gyosida on 4/11/16.
- */
-public class RetrofitFactory {
-
-    private static RetrofitFactory retrofitFactory;
-    private Retrofit retrofit;
-
-    private RetrofitFactory() {
-        init();
-    }
-
-    private void init() {
-        HttpLoggingInterceptor loggingInterceptor = new HttpLoggingInterceptor();
-        loggingInterceptor.setLevel(HttpLoggingInterceptor.Level.BASIC);
-        OkHttpClient client = new OkHttpClient();
-        client.interceptors().add(loggingInterceptor);
-        client.interceptors().add(new Interceptor() {
-            @Override
-            public Response intercept(Chain chain) throws IOException {
-                String token = PreferencesManager.get().getToken();
-                if (token != null) {
-                    Request request = chain.request().newBuilder().addHeader("Authorization", token).build();
-                    return chain.proceed(request);
-                }
-                return chain.proceed(chain.request());
-            }
-        });
-        retrofit = new Retrofit.Builder()
-                .addConverterFactory(GsonConverterFactory.create())
-                .baseUrl("https://allstars-belatrix.herokuapp.com:443/api/")
-                .client(client)
-                .build();
-    }
-
-    public static RetrofitFactory get() {
-        if (retrofitFactory == null) {
-            retrofitFactory = new RetrofitFactory();
-        }
-        return retrofitFactory;
-    }
-
-    public <T> T create(Class<T> clazz) {
-        return retrofit.create(clazz);
-    }
-
-}
